requirements.txt
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
custom-json==1.0.0
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pydantic==1.10.21
pytest==8.3.4
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.3
tokenizers==0.13.3
tomli==2.2.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.30.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.22.0
